+++ CHANGES.rst
@@ -6,7 +6,7 @@ CHANGES
 3.0.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fix imports in `testing.py`
 
 
 3.0.0 (2021-02-22)